utils: split out address-related functions into a new file
and remove unused GetFirstAddr function
